+++ signalfx/resource_signalfx_dashboard_group.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 	dashboard_group "github.com/signalfx/signalfx-go/dashboard_group"
 )
 
@@ -110,13 +111,45 @@ func dashboardGroupResource() *schema.Resource {
 				Type:        schema.TypeSet,
 				Optional:    true,
 				Elem:        &schema.Schema{Type: schema.TypeString},
+				Deprecated:  "Please use permissions field now",
 				Description: "Team IDs that have write access to this dashboard",
 			},
 			"authorized_writer_users": &schema.Schema{
 				Type:        schema.TypeSet,
 				Optional:    true,
 				Elem:        &schema.Schema{Type: schema.TypeString},
+				Deprecated:  "Please use permissions field now",
 				Description: "User IDs that have write access to this dashboard",
+			},
+			"permissions": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Computed:    true,
+				Description: "The custom access control list for this dashboard",
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"principal_id": {
+							Type:        schema.TypeString,
+							Required:    true,
+							Description: "ID of the principal with access",
+						},
+						"principal_type": {
+							Type:         schema.TypeString,
+							Required:     true,
+							ValidateFunc: validation.StringInSlice([]string{"ORG", "TEAM", "USER"}, false),
+							Description:  "Type of principal, possible values: ORG, TEAM, USER",
+						},
+						"actions": {
+							Type:     schema.TypeSet,
+							Optional: true,
+							Elem: &schema.Schema{
+								Type:         schema.TypeString,
+								ValidateFunc: validation.StringInSlice([]string{"READ", "WRITE"}, false),
+							},
+							Description: "Actions level, possible values: READ, WRITE",
+						},
+					},
+				},
 			},
 			"import_qualifier": &schema.Schema{
 				Type:     schema.TypeSet,
@@ -214,6 +247,10 @@ func getPayloadDashboardGroup(d *schema.ResourceData) *dashboard_group.CreateUpd
 			users = append(users, v.(string))
 		}
 		cudgr.AuthorizedWriters.Users = users
+	}
+	permissions := getPermissions(d)
+	if permissions.Acl != nil {
+		cudgr.Permissions = permissions
 	}
 
 	// Because at present the `DashboardConfigs` mirrors the `Dashboards`
@@ -348,6 +385,30 @@ func getPayloadDashboardGroup(d *schema.ResourceData) *dashboard_group.CreateUpd
 	return cudgr
 }
 
+func getPermissions(d *schema.ResourceData) *dashboard_group.ObjectPermissions {
+	permissions := &dashboard_group.ObjectPermissions{}
+	if val := getPermissionsAcl(d); len(val) > 0 {
+		permissions.Acl = val
+	}
+	return permissions
+}
+
+func getPermissionsAcl(d *schema.ResourceData) []*dashboard_group.AclEntry {
+	acl := d.Get("permissions").(*schema.Set).List()
+	aclList := make([]*dashboard_group.AclEntry, len(acl))
+	for i, entry := range acl {
+		entry := entry.(map[string]interface{})
+
+		item := &dashboard_group.AclEntry{
+			PrincipalId:   entry["principal_id"].(string),
+			PrincipalType: entry["principal_type"].(string),
+			Actions:       expandStringSetToSlice(entry["actions"].(*schema.Set)),
+		}
+		aclList[i] = item
+	}
+	return aclList
+}
+
 func dashboardgroupCreate(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*signalfxConfig)
 	payload := getPayloadDashboardGroup(d)
@@ -395,6 +456,23 @@ func dashboardGroupAPIToTF(d *schema.ResourceData, dg *dashboard_group.Dashboard
 				users[i] = v
 			}
 			if err := d.Set("authorized_writer_users", schema.NewSet(schema.HashString, users)); err != nil {
+				return err
+			}
+		}
+	}
+
+	if dg.Permissions != nil {
+		p := dg.Permissions
+		if p.Acl != nil && len(p.Acl) > 0 {
+			acl := make([]map[string]interface{}, len(p.Acl))
+			for i, a := range p.Acl {
+				entry := make(map[string]interface{})
+				entry["principal_id"] = a.PrincipalId
+				entry["principal_type"] = a.PrincipalType
+				entry["actions"] = flattenStringSliceToSet(a.Actions)
+				acl[i] = entry
+			}
+			if err := d.Set("permissions", acl); err != nil {
 				return err
 			}
 		}